feat: make edit drawer mobile responsive
The system is perfect because the width is still hardcoded, but now it is much smaller and more manageable on mobile. I tested this on my iPhone, so results may vary.

Additionally it is not possible to adjust the width on mobile. Perhaps a more dynamic system is needed here.